requirements.txt
Flask==0.9
Jinja2==2.7
MarkupSafe==0.18
Pillow==2.0.0
Werkzeug==0.8.3
distribute==0.6.34
ipython==0.13.2
matplotlib==1.2.1
numpy==1.7.1
wsgiref==0.1.2
